* src/server/sv_game.c changed pf_save to short* * src/game/g_client.c sending lenght of particle name string as short in EV_SPAWN_PARTICLE --- byte was enough, but now the code is more maintainable * src/game/g_main.c sending the lenght of CharacterData as short for EV_RESULTS --- this time we may avoid serious bugs (but I could have introduced new ones) * src/qcommon/common.c changed size of '*' format to short, merging the ssianky 'n' hack from r1807 with the old '*' format; corrected MSG_LengthFormat for '!' that do not occur at the format start (no such occurences right now) * src/client/cl_team.c reading CharacterData with size given as short, now byte * src/client/cl_parse.c corrected a grave bug for EV_ACTOR_SHOOT_HIDDEN, where MSG_ReadByte appeared inside GET_FIREDEF macro and got doubled --- I will check if there are any other such errors in the code; set up a check for not read correctly EV_* messages and eliminated many false positives; corrected the bug with 2 first bytes of EV_INV_ADD message missed

+++ src/game/g_client.c
@@ -1558,7 +1558,7 @@ static void G_Damage(edict_t * ent, int dmgtype, int damage, edict_t * attacker)
 				gi.AddEvent(PM_ALL, EV_SPAWN_PARTICLE);
 				gi.WriteShort(ent->spawnflags);
 				gi.WriteGPos(ent->pos);
-				gi.WriteByte((int)strlen(ent->particle));
+				gi.WriteShort((int)strlen(ent->particle));
 				gi.WriteString(ent->particle);
 			}
 			gi.unlinkentity(ent);
